homework7 fixed, ready for second review
